Added Edge loading function

+++ app/src/main/java/graph/Vertex.java
@@ -2,9 +2,9 @@ package graph;
 
 public class Vertex {
     int id;
-    int x;
-    int y;
-    public Vertex(int id, int x, int y) {
+    float x;
+    float y;
+    public Vertex(int id, float x, float y) {
         this.id = id;
         this.x = x;
         this.y = y;
